[Notebook] Remove command for "Execute All"
Because F6 is too close to F5 and it 's dangerous when certain cells are heavy-duty and we accidentally click on F6 (which happened a few times). Besides, for such "one-shot" action, having a menu button is not just safer but also makes perfect sense - sort of serve as a "confirmation".