[Core] Enhance unit test to point out limits of `Parse()`
- This particular issue only happens when `Parse()` is used to run scripts that also calls `Parse()`, which will then mess up with state.

This situation generally doesn't happen to Notebook or Pure REPL, because they ARE the hosting environment and the entered scripts are passed to Parse() directly.
But this happens when using CentralSnippets as of current implementation (aka. CentralSnippets calls `Parse()`, but itself is invoked inside a `Parse()` context, either from Notebook or Pure REPL).

(This implies that CentralSnippets must be implemented on the language level as macros or somehow we must workaround this limit in RoslynContext.Parse())